+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/manager/EventCatchupProcessor.java
@@ -4,10 +4,7 @@ import static java.lang.String.format;
 
 import uk.gov.justice.services.event.source.subscriptions.repository.jdbc.SubscriptionsRepository;
 import uk.gov.justice.services.eventsourcing.source.core.EventSource;
-import uk.gov.justice.services.messaging.JsonEnvelope;
 import uk.gov.justice.subscription.domain.subscriptiondescriptor.Subscription;
-
-import java.util.stream.Stream;
 
 import org.slf4j.Logger;
 
@@ -16,19 +13,19 @@ public class EventCatchupProcessor {
     private final Subscription subscription;
     private final SubscriptionsRepository subscriptionsRepository;
     private final EventSource eventSource;
-    private final EventBufferProcessor eventBufferProcessor;
+    private final TransactionalEventProcessor transactionalEventProcessor;
     private final Logger logger;
 
     public EventCatchupProcessor(
             final Subscription subscription,
             final EventSource eventSource,
-            final EventBufferProcessor eventBufferProcessor,
             final SubscriptionsRepository subscriptionsRepository,
+            final TransactionalEventProcessor transactionalEventProcessor,
             final Logger logger) {
         this.subscription = subscription;
         this.subscriptionsRepository = subscriptionsRepository;
         this.eventSource = eventSource;
-        this.eventBufferProcessor = eventBufferProcessor;
+        this.transactionalEventProcessor = transactionalEventProcessor;
         this.logger = logger;
     }
 
@@ -39,15 +36,12 @@ public class EventCatchupProcessor {
         final long eventNumber = subscriptionsRepository.getOrInitialiseCurrentEventNumber(subscription.getName());
 
         final int totalEventsProcessed = eventSource.findEventsSince(eventNumber)
-                .mapToInt(this::process)
+                .mapToInt(transactionalEventProcessor::processWithEventBuffer)
                 .sum();
 
         logger.info(format("Event catchup retrieved and processed %d new events", totalEventsProcessed));
         logger.info("Event catchup complete");
     }
 
-    private int process(final JsonEnvelope event) {
-        eventBufferProcessor.processWithEventBuffer(event);
-        return 1;
-    }
+
 }