+++ event-sourcing/subscription-manager/src/main/java/uk/gov/justice/services/event/sourcing/subscription/manager/cdi/factories/EventCatchupProcessorFactory.java
@@ -5,6 +5,7 @@ import static org.slf4j.LoggerFactory.getLogger;
 import uk.gov.justice.services.event.source.subscriptions.repository.jdbc.SubscriptionsRepository;
 import uk.gov.justice.services.event.sourcing.subscription.manager.EventBufferProcessor;
 import uk.gov.justice.services.event.sourcing.subscription.manager.EventCatchupProcessor;
+import uk.gov.justice.services.event.sourcing.subscription.manager.TransactionalEventProcessor;
 import uk.gov.justice.services.eventsourcing.source.core.EventSource;
 import uk.gov.justice.subscription.domain.subscriptiondescriptor.Subscription;
 
@@ -17,12 +18,18 @@ public class EventCatchupProcessorFactory {
     @Inject
     SubscriptionsRepository subscriptionsRepository;
 
-    public EventCatchupProcessor create(final Subscription subscription, final EventSource eventSource, final EventBufferProcessor eventBufferProcessor) {
+    public EventCatchupProcessor create(
+            final Subscription subscription,
+            final EventSource eventSource,
+            final EventBufferProcessor eventBufferProcessor) {
+
+        final TransactionalEventProcessor transactionalEventProcessor = new TransactionalEventProcessor(eventBufferProcessor);
+
         return new EventCatchupProcessor(
                 subscription,
                 eventSource,
-                eventBufferProcessor,
                 subscriptionsRepository,
+                transactionalEventProcessor,
                 getLogger(EventCatchupProcessor.class)
         );
     }